Move Event to event-details
