+++ event-details/Event.ts
