+++ lib/events/index.ts
@@ -1,2 +1,2 @@
-export * from "./Event"
+export * from "../../event-details/Event"
 export * from "./SaveEvent"
